+++ render/sdl/sdl_scancode.mojo
@@ -51,7 +51,7 @@ struct Scancode(Indexer, Intable, TrivialRegisterPassable):
 
     @always_inline
     fn __init__(out self, value: Int):
-        self.value = value
+        self.value = UInt32(value)
 
     @always_inline
     fn __int__(self) -> Int:
